refactor: Add token blacklist functionality in server.js
This commit adds functionality to blacklist tokens in the `verifyToken` middleware function in the `server.js` file. It introduces a new array `blacklistSessions` to store revoked tokens and checks if a token is in the blacklist before verifying it. If a token is found in the blacklist, a 401 status with a "Token has been revoked" message is returned. This enhancement improves the security of the authentication process.