feat: Add endpoint to fetch unfinished customer courses
This commit adds a new endpoint `/getFinishedCourse` to the `server.js` file. The endpoint is protected by token verification and fetches customer courses that have been marked as finished from the `tcustomer_course` table in the database. The fetched courses are returned in the response as an array. This change enhances the functionality of the server by providing a way to retrieve finished customer courses.