refactor: Fix SQL syntax error in finishCustomerCourse endpoint
This commit fixes a SQL syntax error in the server.js file. The error was caused by an extra closing parenthesis in the UPDATE query of the finishCustomerCourse endpoint. The extra parenthesis has been removed, ensuring the query executes correctly. This change improves the reliability and functionality of the endpoint.